Update README.md
In this code made to run in a website, you can add numbers between 1 and 100 in a list, without repetead ones, and click in the button in the end to show the quantity of numbers added, the smallest and the biggest number in the list, the result of the addition and the average of the numbers added on that list. Made with HTML, CSS and Javascript.